+++ multi/client.go
@@ -1,4 +1,4 @@
-// Copyright © 2021, 2022 Attestant Limited.
+// Copyright © 2021 - 2024 Attestant Limited.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -20,11 +20,12 @@ import (
 
 	consensusclient "github.com/attestantio/go-eth2-client"
 	"github.com/attestantio/go-eth2-client/api"
+	"github.com/attestantio/go-eth2-client/http"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
 )
 
-// monitor monitors active and inactive connections, and moves them between
+// monitor monitors active and inactive clients, and moves them between
 // lists accordingly.
 func (s *Service) monitor(ctx context.Context) {
 	log := s.log.With().Logger()
@@ -38,13 +39,13 @@ func (s *Service) monitor(ctx context.Context) {
 
 			return
 		case <-time.After(30 * time.Second):
-			s.recheck(ctx)
+			s.recheck(ctx, false)
 		}
 	}
 }
 
 // recheck checks clients to update their state.
-func (s *Service) recheck(ctx context.Context) {
+func (s *Service) recheck(ctx context.Context, forceUpdate bool) {
 	// Fetch all clients.
 	clients := make([]consensusclient.Service, 0, len(s.activeClients)+len(s.inactiveClients))
 	s.clientsMu.RLock()
@@ -54,15 +55,21 @@ func (s *Service) recheck(ctx context.Context) {
 
 	// Ping each client to update its state.
 	for _, client := range clients {
-		if ping(ctx, client) {
+		if forceUpdate {
+			if httpClient, isHTTPClient := client.(*http.Service); isHTTPClient {
+				httpClient.CheckConnectionState(ctx)
+			}
+		}
+		switch {
+		case client.IsActive():
 			s.activateClient(ctx, client)
-		} else {
+		default:
 			s.deactivateClient(ctx, client)
 		}
 	}
 }
 
-// deactivateClient deactivates a client, moving it to the inactive list if not currently on it.
+// deactivateClient marks a client as deactivated, moving it to the inactive list if not currently on it.
 func (s *Service) deactivateClient(ctx context.Context, client consensusclient.Service) {
 	log := zerolog.Ctx(ctx)
 
@@ -74,19 +81,21 @@ func (s *Service) deactivateClient(ctx context.Context, client consensusclient.S
 	for _, activeClient := range s.activeClients {
 		if activeClient == client {
 			inactiveClients = append(inactiveClients, activeClient)
-			setProviderActiveMetric(ctx, client.Address(), "inactive")
+			setProviderStateMetric(ctx, client.Address(), "inactive")
 		} else {
 			activeClients = append(activeClients, activeClient)
 		}
 	}
 	if len(inactiveClients) != len(s.inactiveClients) {
-		log.Trace().Str("client", client.Address()).Int("active", len(activeClients)).Int("inactive", len(inactiveClients)).Msg("Client deactivated")
+		log.Trace().Str("client", client.Address()).
+			Int("active", len(activeClients)).
+			Int("inactive", len(inactiveClients)).
+			Msg("Client deactivated")
 	}
 
 	s.activeClients = activeClients
-	setProvidersMetric(ctx, "active", len(s.activeClients))
 	s.inactiveClients = inactiveClients
-	setProvidersMetric(ctx, "inactive", len(s.inactiveClients))
+	setConnectionsMetric(ctx, len(s.activeClients), len(s.inactiveClients))
 }
 
 // activateClient activates a client, moving it to the active list if not currently on it.
@@ -101,41 +110,21 @@ func (s *Service) activateClient(ctx context.Context, client consensusclient.Ser
 	for _, inactiveClient := range s.inactiveClients {
 		if inactiveClient == client {
 			activeClients = append(activeClients, inactiveClient)
-			setProviderActiveMetric(ctx, client.Address(), "active")
+			setProviderStateMetric(ctx, client.Address(), "active")
 		} else {
 			inactiveClients = append(inactiveClients, inactiveClient)
 		}
 	}
 	if len(inactiveClients) != len(s.inactiveClients) {
-		log.Trace().Str("client", client.Address()).Int("active", len(activeClients)).Int("inactive", len(inactiveClients)).Msg("Client activated")
+		log.Trace().Str("client", client.Address()).
+			Int("active", len(activeClients)).
+			Int("inactive", len(inactiveClients)).
+			Msg("Client activated")
 	}
 
 	s.activeClients = activeClients
-	setProvidersMetric(ctx, "active", len(s.activeClients))
 	s.inactiveClients = inactiveClients
-	setProvidersMetric(ctx, "inactive", len(s.inactiveClients))
-}
-
-// ping pings a client, returning true if it is ready to serve requests and
-// false otherwise.
-func ping(ctx context.Context, client consensusclient.Service) bool {
-	log := zerolog.Ctx(ctx)
-
-	provider, isProvider := client.(consensusclient.NodeSyncingProvider)
-	if !isProvider {
-		log.Debug().Str("provider", client.Address()).Msg("Client does not provide sync state")
-
-		return false
-	}
-
-	response, err := provider.NodeSyncing(ctx, &api.NodeSyncingOpts{})
-	if err != nil {
-		log.Warn().Err(err).Msg("Failed to obtain sync state from node")
-
-		return false
-	}
-
-	return (!response.Data.IsSyncing) || (response.Data.HeadSlot == 0 && response.Data.SyncDistance == 0)
+	setConnectionsMetric(ctx, len(s.activeClients), len(s.inactiveClients))
 }
 
 // callFunc is the definition for a call function.  It provides a generic return interface
@@ -159,14 +148,14 @@ func (s *Service) doCall(ctx context.Context, call callFunc, errHandler errHandl
 
 	if len(activeClients) == 0 {
 		// There are no active clients; attempt to re-enable the inactive clients.
-		s.recheck(ctx)
+		s.recheck(ctx, true)
 		s.clientsMu.RLock()
 		activeClients = s.activeClients
 		s.clientsMu.RUnlock()
 	}
 
 	if len(activeClients) == 0 {
-		return nil, errors.New("no active clients to which to make call")
+		return nil, errors.New("no clients to which to make call")
 	}
 
 	var err error
@@ -174,15 +163,19 @@ func (s *Service) doCall(ctx context.Context, call callFunc, errHandler errHandl
 	for _, client := range activeClients {
 		res, err = call(ctx, client)
 		if err != nil {
-			log.Trace().Err(err).Msg("Potentially deactivating client due to error")
+			log.Trace().Str("client", client.Name()).Str("address", client.Address()).Err(err).Msg("Potentially deactivating client due to error")
 			var apiErr *api.Error
-			if errors.As(err, &apiErr) && apiErr.StatusCode/100 == 4 {
+			switch {
+			case errors.As(err, &apiErr) && apiErr.StatusCode/100 == 4:
 				log.Trace().Str("client", client.Name()).Str("address", client.Address()).Err(err).Msg("Not deactivating client on user error")
 
 				return res, err
-			}
-			if errors.Is(err, context.Canceled) {
+			case errors.Is(err, context.Canceled):
 				log.Trace().Str("client", client.Name()).Str("address", client.Address()).Msg("Not deactivating client on canceled context")
+
+				return res, err
+			case errors.Is(err, context.DeadlineExceeded):
+				log.Trace().Str("client", client.Name()).Str("address", client.Address()).Msg("Not deactivating client on context deadline exceeded")
 
 				return res, err
 			}
@@ -191,10 +184,8 @@ func (s *Service) doCall(ctx context.Context, call callFunc, errHandler errHandl
 			if errHandler != nil {
 				failover, err = errHandler(ctx, client, err)
 			}
-
 			if failover {
 				log.Debug().Str("client", client.Name()).Str("address", client.Address()).Err(err).Msg("Deactivating client on error")
-				// Failed with this client; try the next.
 				s.deactivateClient(ctx, client)
 
 				continue
@@ -227,6 +218,8 @@ func (s *Service) providerInfo(ctx context.Context, provider consensusclient.Ser
 			switch {
 			case strings.Contains(strings.ToLower(response.Data), "lighthouse"):
 				providerName = "lighthouse"
+			case strings.Contains(strings.ToLower(response.Data), "lodestar"):
+				providerName = "lodestar"
 			case strings.Contains(strings.ToLower(response.Data), "prysm"):
 				providerName = "prysm"
 			case strings.Contains(strings.ToLower(response.Data), "teku"):